refactor: extract browser-extension external CDP accessor into lib module
Phase 1 of content.ts (7,311L→7,229L) extraction. Extracted
installExternalAccessor + getExternalAccessorStatus into
lib/external-accessor.ts (159L) using dependency-injection pattern.
All 70 browser-extension tests pass, make check 0 errors.
